fix: put request in util/requests
