+++ src/components/app/FileMenu.js
@@ -8,8 +8,7 @@ import { useD2 } from '@dhis2/app-runtime-adapter-d2';
 import { useDataMutation, useDataEngine } from '@dhis2/app-runtime';
 import { newMap, tOpenMap, setMapProps } from '../../actions/map';
 import { setAlert } from '../../actions/alerts';
-import { dataStatisticsMutation } from '../../util/apiDataStatistics.js';
-import { fetchMap } from '../../util/requests';
+import { fetchMap, dataStatisticsMutation } from '../../util/requests';
 import { cleanMapConfig } from '../../util/favorites';
 import { useSystemSettings } from '../SystemSettingsProvider';
