--- src/util/apiDataStatisitcs.js
@@ -1,8 +1,0 @@
-export const dataStatisticsMutation = {
-    resource: 'dataStatistics',
-    params: ({ id }) => ({
-        favorite: id,
-        eventType: 'MAP_VIEW',
-    }),
-    type: 'create',
-};
